Add user security restrictions to interface CRUD operations
Add checks to ensure the http request being sent is done by the owner of the object, and not a malicious user. A test case covering unauthorized Note operations is also provided.

+++ src/main/java/com/udacity/jwdnd/course1/cloudstorage/controller/HomeController.java
@@ -5,11 +5,14 @@ import com.udacity.jwdnd.course1.cloudstorage.entity.File;
 import com.udacity.jwdnd.course1.cloudstorage.entity.Note;
 import com.udacity.jwdnd.course1.cloudstorage.service.*;
 import org.springframework.http.*;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.io.FileNotFoundException;
 
 @Controller
 @RequestMapping("/home")
@@ -30,10 +33,12 @@ public class HomeController {
     }
 
     @GetMapping
-    public String homeView(@ModelAttribute Note note, Model model) {
-        model.addAttribute("files", fileService.getAllFiles());
-        model.addAttribute("notes", noteService.getAllNotes());
-        model.addAttribute("credentials", credentialService.getAllCredentials());
+    public String homeView(Authentication authentication, @ModelAttribute Note note, Model model) {
+        Integer userId = userService.getUser(authentication.getName()).getUserId();
+
+        model.addAttribute("files", fileService.getAllFiles(userId));
+        model.addAttribute("notes", noteService.getAllNotes(userId));
+        model.addAttribute("credentials", credentialService.getAllCredentials(userId));
         model.addAttribute("encryptionService", encryptionService);
 
         return "home";
@@ -71,13 +76,21 @@ public class HomeController {
     @GetMapping(value = "/file", params = "view")
     @ResponseBody
     public ResponseEntity<byte[]> viewFile(@RequestParam("view") Integer fileId,
+                                           Authentication authentication,
                                            Model model) {
+        Integer userId = userService.getUser(authentication.getName()).getUserId();
+
         HttpHeaders headers = new HttpHeaders();
         headers.setCacheControl(CacheControl.noCache().getHeaderValue());
 
         File file = fileService.getFile(fileId);
         if (file == null) {
             model.addAttribute("fileError", "Ops! Error retrieving file. Please, try again!");
+            return ResponseEntity.badRequest().body(null);
+        }
+
+        if (!file.getUserId().equals(userId)) {
+            throw new AccessDeniedException("You do not have access for this operation");
         }
 
         headers.setContentType(MediaType.parseMediaType(file.getContentType()));
@@ -89,10 +102,17 @@ public class HomeController {
     }
 
     @GetMapping(value = "/file", params = "delete")
-    public String deleteFile(@RequestParam("delete") Integer fileId, Model model) {
+    public String deleteFile(@RequestParam("delete") Integer fileId,
+                             Authentication authentication,
+                             Model model) {
+        Integer userId = userService.getUser(authentication.getName()).getUserId();
         String deleteSuccess = null;
         String deleteError = null;
 
+        Integer fileUserId = fileService.getFile(fileId).getUserId();
+        if (userId != fileUserId) {
+            throw new AccessDeniedException("You do not have access for this operation");
+        }
         int rowsDeleted = fileService.deleteFile(fileId);
         if (rowsDeleted < 1) {
             deleteError = "Ops! Could not delete file. Please, try again!";
@@ -105,7 +125,7 @@ public class HomeController {
         } else {
             model.addAttribute("fileError", deleteError);
         }
-        model.addAttribute("files", fileService.getAllFiles());
+        //model.addAttribute("files", fileService.getAllFiles());
 
         return "redirect:/home";
     }
@@ -148,13 +168,21 @@ public class HomeController {
             model.addAttribute("noteError", noteError);
         }
 
-        // TODO - attempt to redirect to a fragment - the tab does not change, however
-        return "redirect:/home#nav-notes";
+        return "redirect:/home";
     }
 
     @GetMapping(value = "/note", params = "delete")
-    public String deleteNote(@RequestParam("delete") Integer noteId, Model model) {
+    public String deleteNote(@RequestParam("delete") Integer noteId,
+                             Authentication authentication,
+                             Model model) {
+        Integer userId = userService.getUser(authentication.getName()).getUserId();
         String deleteError = null;
+
+        Integer noteUserId = noteService.getNote(noteId).getUserId();
+        if (userId != noteUserId) {
+            throw new AccessDeniedException("You do not have access for this operation");
+        }
+
         int rowsDeleted = noteService.deleteNote(noteId);
         if (rowsDeleted < 1) {
             deleteError = "Ops! Could not delete note. Please, try again!";
@@ -206,8 +234,18 @@ public class HomeController {
     }
 
     @PostMapping("/delete-credential")
-    public String deleteCredential(@ModelAttribute Credential credential, Model model) {
+    public String deleteCredential(@ModelAttribute Credential credential,
+                                   Authentication authentication,
+                                   Model model) {
+        Integer userId = userService.getUser(authentication.getName()).getUserId();
         String deleteError = null;
+
+        Integer credentialUserId = credentialService
+                .getCredential(credential.getCredentialId())
+                .getUserId();
+        if (userId != credentialUserId) {
+            throw new AccessDeniedException("You do not have access for this operation");
+        }
         int rowsDeleted = credentialService.deleteCredential(credential.getCredentialId());
         if (rowsDeleted < 1) {
             deleteError = "Oh snap! Could not delete the credential! Please, try again!";
